requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
docker==5.0.3
executing==0.8.2
idna==3.3
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
requests==2.28.0
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.3
wllvm==1.3.1
libclang==14.0.6
tomli==2.0.1
numpy==1.24.4
tabulate==0.9.0
graphviz==0.20.1
networkx==3.1
matplotlib==3.7.3
flask==3.0.3
Levenshtein==0.25.1
scikit-learn==1.3.2